Add build images CI job
Adds a new GitHub workflow that builds and pushes the operators images
(manager/bundle/catalog) to the kuadrant quay.io repo.
Updates the Makefile to bring it inline with the rest of the kuadrant
operators allowing the same GH workflow to be used in all kuadrant
operator repos.
Adds authorino and deploy kustomizations.